Added support for activity starting from any component.
Summary:
Using the @ActivityElement and @UsesActivities annotations, it is now
possible to launch an arbitrary activity from a component without
hacking Compiler.java to register that activity in AndroidManifest.xml.

Using the @ReceiverElement and @UsesBroadcastReceivers annotations, it is
now possible to leverage the full functionality of the Android
framework when creating <receiver> elements in AndroidManifest.xml.

Both of these new annotations allow App Inventor component developers
to add <intent-filter> and <meta-data> subelements to activities and
broadcast receivers declared in AndroidManifest.xml with the
@IntentFilterElement and @MetaDataElement annotations.

-ListPicker has been refactored to use these new annotations.
-BarcodeScanner has been refactored to use these new annotations.
-Twitter has been refactored to use these new annotations.
-Texting has been refactored to use these new annotations.
-Since the NearField component makes changes to the main activity contingent
 on information that is only available at build time, Form and NearField
 cannot be refactored to use these annotations at the moment.
-New unit tests for broadcast receiver and activity generation were added in
 com.google.appinventor.buildserver.CompilerTest.java.
-Added legacy support for the deprecated @SimpleBroadcastReceiver annotation
 so that we don't break extensions that are currently using it.

Change-Id: I2114f2b522bdbcf51e1e1b79a642010ff040e183