Do not erase UserSettings when App Inventor Closed
There is a race condition where if App Inventor is closed (browser
window closed) while the system is still initializing, the user’s global
settings will be over-written by an empty settings object because the
global settings have not finished loading. We detect that with this code
and avoid over-writing the settings.

Change-Id: Icfcf8204b860984bf36e2a7d525ebf5814a2ffd8