Update settings on server only if changes occur
Switching between tabs will save settings even if nothing has
changed. This results in more network operations than is necessary to
maintain correct behavior of the application.

This commit updates the CommonSettings class hierarchy to record when
changes have been made. UserSettings and ProjectSettings will perform
RPC if and only if the changed flag is true. Otherwise, saving
settings is a no-op.

Change-Id: I1f15d39418036b320ee3050003034b97e46415be