Cache Flag and GallerySetting state
The loading of GallerySettings has a high latency in production. This
commit makes it so that the GallerySetting object is created only once
per GAE instance. It also updates the Flag class to cache the value on
first read since these flags are set in appengine-web.xml and are only
expected to change during deployment of a new version of App Inventor.

Change-Id: I7bb123697e821f2510807c5cbd5671bae71db307