Permit newer version projects to run
Remove the check in versioning.js that throws an exception if a
component version is newer then the system version. This exception
guaranteed that a project touched by a newer version of App Inventor
would not work in the current version.

With this change the project may work if the newer version made not
incompatible changes. The user is shown a warning dialog stating that
the project might not work.

Change-Id: I72c15383b20d918ffd0077d6f7a10f8d1dd4b345