Fix vertical parameter orientation on procedure blocks
Procedure definitions have the option for horizontal/vertical
orientation of parameters. However, switching resulted in the inputs
disappearing from the block. This was caused by the new fields not
being initialized correctly such that when the block attempted to
redraw, expected SVG elements were undefined.

This commit fixes the updateParams_ method for procedure blocks so
that they will correctly handle switching between the two cases.

Change-Id: Ib9a161548954a2990a0351ad07a808c35d2bc928