Fix variable reference renaming
When renaming variables, references to those variables are not
correctly renamed. This is due to a change in the order in which
Blockly manipulates the field's value and calls the validator. Since
the update occurs before validation, we don't know the original name
and therefore cannot find and update all of the references. This
commit provides the flydown with its own implementation of
onHtmlInputChanged_ so that it can provide a behavior that is
compatible with our renaming mechanism, mainly that it will validate
first and only if successful will it update the field's value.

Change-Id: I7569d0f0b1b6f57e476df5e32f1b2a12e7d7533c