Add DomHelper for goog.ui.Dialog
Because Blockly is linked into the GWT JavaScript, goog.ui.Dialog
attempts to show itself on the GWT iframe's document rather than on
the top window where the UI is actually displayed. This commit adds a
goog.ui.DomHelper so that the Dialog attachs its DOM elements to the
correct document.

NB: This has the side-effect that the runtime error dialog from the
companion is now UI global (it will show even if in the designer).

Change-Id: I3f7fb60d8898e37a689cbc3f406ef39832717d99