Fix buildserver paths on Windows
The buildserver should return an error if Kawa cannot compile the YAIL
file. However, on Windows we end up always pointing at the classes
directory, which always exists, and therefore we continue compiling an
app even though the class files are missing. On Mac/Linux the same
project results in a Kawa compilation error. The logic error is
introduced by the hard coding of the path "/../src/" in
Compiler.java. On Windows the path uses the conventional Windows
separator "\\", so the substring of interest is "\\..\\src\\". This
commit adjusts the lookup string to use File.separator so that the
lookup path will be correct regardless of platform.

Change-Id: Id066cad6710e7fe8c6710b40465c87992f547854