Fix issue with deleting an instantiated extension
It was reported that deleting an extension that had been instantiated
results in an "Oops! something broke" message. The issue is caused by
collecting all of the sibling component types in the same package (due
to extension bundles) and accidentally double-counting the extension
type that triggered the deletion. This would result in an attempt to
delete the same component twice resulting in an error because the
component no longer existed.

This commit fixes the problem by checking the sibling name against the
set of to-be-removed classes to ensure that adding a sibling doesn't
duplicate the original class for removal.

Change-Id: Id3d5e17da4e3a5c4e8423ac32fc427cfdcd8ebdc