Eliminate IllegalArgumentExceptions
If someone attempts to download a packaged project while a new build is
running, we throw in IllegalArgumentException because the built apk is
no longer present (the first thing a new build does is remove the old
apk). This happens often enough that it is one of the top errors being
logged by the Google Cloud Console.

This change eliminates the IllegalArgumentException and instead just
returns a “File Not Found” error (404 error) to the user attempting the
download.

Change-Id: I6fb92e51477c1ad93059cf57a449396d4fea0d7d