Add option to generate extensions with old-style layout
Prior to the IOT release, extensions were packaged using the FQCN of
the extension inside of the AIX file. To faciltate being able to
contain multiple extension components in a single package, we switched
to using the package name rather than the FQCN. This causes backward
compatibility issues for extension developers who released extensions
prior to the update.

This commit adds a extensions.packagefqcn property to ant that can be
set to true by developers supporting older extensions so that the
internals of the extension package reflect the older layout. It is
recommended that all new extensions be built with the default (false)
to benefit from the new layout.

Change-Id: Idd9134c7d4d7c19f9a8869e4f88e164e2efa5cc8