Fix cast exception from ScaleDrawable to ClipDrawble
We grab the progress drawable and cast it to ClipDrawable. This causes
a ClassCastException on phones where the manufacturer provides custom
view hierarchies for the slider but do not use the same classes as the
original Android views. However, the only method we call is defined on
Drawable, so this commit switches to using the Drawable superclass and
removes the cast to ClipDrawable.

Fixes #912

Change-Id: I20b0691ca684a2e34140aafa9d3b04eb4adf7148