Make tests run from module directory
Before this commit, tests were run with the current working directory
different depending on whether the tests were run from within the
submodule directory (e.g., components) versus in the top level ant
script. This means that relative paths may not be correct if one wants
to run tests only on a specific submodule.

This commit makes it so that the ai.dojunit macro changes directory
into the appropriate submodule before running any tests so that
relative paths do not need to be adjusted based on whether all or some
subset of tests are being run.

Change-Id: I3d3c5803f3c29adc9e371e106aed08d92968eb35