Display the Login page instead of empty dialog
During development it is common to erase the local development datastore
and start over. However if the browser in use still had a valid
authentication cookie (very common), a blank dialog would be displayed
because the cookie resolved to a non-existent account. This change
results instead with either the login page, or Google authentication
being invoked (depending on configuration). Note: This code should *not*
get triggered in production.

Change-Id: Iafbc7c774c392d99073342c49108dc6774e715a0